+++ sdk/monitor/OpenTelemetry.Exporter.AzureMonitor/src/GeneratedFix/Base.cs
