+++ sdk/monitor/OpenTelemetry.Exporter.AzureMonitor/src/GeneratedFix/Domain.cs
