+++ sdk/monitor/OpenTelemetry.Exporter.AzureMonitor/src/GeneratedFix/Readme.md
@@ -1,0 +1,2 @@
+﻿This directory contains temporary partial classes to fix the auto-generated classes from swagger.
+This directory can be deleted when the fixes are incorporated into the swagger.
