+++ sdk/monitor/OpenTelemetry.Exporter.AzureMonitor/src/GeneratedFix/RequestData.cs
