+++ sdk/monitor/OpenTelemetry.Exporter.AzureMonitor/src/GeneratedFix/StackFrame.cs
