+++ sdk/monitor/OpenTelemetry.Exporter.AzureMonitor/tests/OpenTelemetry.Exporter.AzureMonitor.Demo.Tracing/DemoTrace.cs
@@ -3,6 +3,8 @@
 
 using System.Diagnostics;
 
+using OpenTelemetry.Exporter.AzureMonitor.Extensions;
+
 namespace OpenTelemetry.Exporter.AzureMonitor.Demo.Tracing
 {
     public static class DemoTrace
